allow working directory to be beneath /tmp
When trying to run squilt in a package checkout in /tmp then the
invocation fails, because the mount of the working directory in the
target namespace is no longer visible due to the /tmp directory being
replaced by a private tmp mount.

By moving the directive for the working directory to the end of the
config file things seem to work out correctly again.